Create README.md
Our project is a sql-python connectivity program through which users can run the program as an admin or as a customer. When the program is run as admin, the user can take numerous actions like viewing sales, viewing the stock and restocking it and also add and deleting ingredients. When a user tries to access as admin, they are asked for appropriate usernames and are provided with three tries to enter the right password. As a customer the user can view the menu, add items to cart and view cart and finally request for a bill. Certain items allow further customisation through which the customer can specify the ingredients they desire. When the program is run as customer, they are required to enter their phone number and name after which their details and number of orders are tracked. A discount is provided when the number of orders cross specific values.